refactor: follow test name
